ZOOKEEPER-3399: Remove logging in getGlobalOutstandingLimit for optimal performance.
Please refer https://issues.apache.org/jira/browse/ZOOKEEPER-3399 for detailed description. This PR removes the LOG statement which was the state of the code path before ZOOKEEPER-3177. The original patch of ZOOKEEPER-3177 had the LOG statement in right place, but that was changed in the second pull request (711) that reverted part of the original patch due to a find bug warning.

Author: Michael Han <lhan@twitter.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>

Closes #956 from hanm/ZOOKEEPER-3399