ZOOKEEPER-3391: Drop unused RecFormat(CSV/XML)
As described in

http://zookeeper-user.578899.n2.nabble.com/Deprecated-CSVInputArchive-and-XMLInputArchive-td7584086.html

these 2 input archives(CSV/XML) are not actively maintained and we probably don't have test coverage for them either, so keeping them in the codebase could be questionable.

Without input archives, output archives are also meaningless. Thus transitive removals of `CSVOutputArchive` and `XMLOutputArchive`.

One problem is we also use `CSVOutputArchive` for record `toString` method. For backward compatibility, I rename it to `ToStringOutputArchive` and remain almost all methods. We might discuss another `toString` gen path but it should work as now.

Author: tison <wander4096@gmail.com>

Reviewers: eolivelli@apache.org, andor@apache.org

Closes #983 from TisonKun/ZOOKEEPER-3391