ZOOKEEPER-3548: Redundant zxid check in SnapStream.isValidSnapshot.
`Util.getZxidFromName` is called twice in `SnapStream.isValidSnapshot`, and the second call is redundant: there is no way the check would return false when the second call of `Util.getZxidFromName` is executed as otherwise the code would already return false in the first call of `Util.getZxidFromName`.

No test as the logic is fairly obvious (and the logic is not well testable, in first place).

Author: Michael Han <hanm@apache.org>

Reviewers: Norbert Kalmar <nkalmar@apache.org>

Closes #1092 from hanm/wip