ZOOKEEPER-2282: C Client: chroot not stripped in asynchronous callbacks
This is related to ZOOKEEPER-1027, which adjusted the result of `create` calls to account for the chroot path.  It did not, however, correct the problem for asynchronous calls:

    $ cli_st $ENSEMBLE_HOSTS/rt-46444

    create /x/fred
    Creating [/x/fred] node (mode: 0)
    [/x/fred]: rc = 0
            name = /rt-46444/x/fred

    create2 /x/wilma
    Creating [/x/wilma] node (mode: 0)
    [/x/wilma]: rc = 0
            name = /rt-46444/x/wilma
            ctime = Wed Aug 28 18:14:16 2019
            czxid=1f00000041
            mtime=Wed Aug 28 18:14:16 2019
            mzxid=1f00000041
            version=0	aversion=0
            ephemeralOwner = 0

(It turns out that there was already a ticket, with a patch, for this issue.  I am mentioning Andrew Grasso as a co-author.  This commit also avoids the test failure, however.)

Co-authored-by: Andrew Grasso <andrew-grassousers.noreply.github.com>

Author: Damien Diederen <dd@crosstwine.com>

Reviewers: eolivelli@apache.org, andor@apache.org

Closes #1066 from ztzg/ZOOKEEPER-2282-c-client-async-chroot-paths