ZOOKEEPER-3703: publish a test JAR
It would be very helpful to Apache Curator and others if ZooKeeper published its testing code as a Maven Test JAR. Curator, for example, could use it to improve its testing server to make it easier to inject error conditions without having to have forced time delays and other hacks.

NOTE: if we move forward with gRPC (ZOOKEEPER-102) that would be in a new module and this would be required. So, might as well do it now.

Author: Jordan Zimmerman <jordan@jordjzimmerman@grubhub.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>, Norbert Kalmar <nkalmar@apache.org>

Closes #1229 from Randgalt/ZOOKEEPER-3703-publish-test-jar